# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Mercury Layer BIP448 prototype API
  version: 0.1.0
  description: |
    This contract describes the functional Rocket routes mounted by the
    Mercury server and token server in this repository.

    Cooperative duplicate-sweep boundaries:
    - The client retains one logical Coin and canonical amount. Normalized
      bindings expose every same-script value separately through nested
      coin.duplicates with stable wallet-database-local indices. Passive
      discovery does not sign.
    - Exact CLI command bip448-sweep-duplicate selects one confirmed duplicate
      and creates one one-input/one-output transaction. Its checked fee is
      ceil(112 * fee_rate_sat_per_vbyte), and a dust result is rejected before
      signing. Each output has its own transaction, fee, and signing count.
      Dust and unresolved bindings remain visible and can block canonical
      close.
    - Attempt artifacts and phases are durable and exact retry reuses them.
      SecondArmed is persisted before a possibly delivered sign/second and
      permanently makes the statechain exit-only. Duplicate sweeps never
      delete statechain server rows; canonical close is last and requires every
      known current-owner duplicate to be handled.
    - The client durably stores a transfer intent before sender mutation. With
      no different authenticated request intervening, exact same-request
      /transfer/sender response-loss replay returns one stored x1 for the
...
servers:
...
paths:
...
  /transfer/sender:
    post:
      tags: [Transfer]
      summary: Create server transfer state for a recipient
      description: |
        Authentication, owner-generation checks, batch validation, replay, and
        insertion use the locked current-owner row. With no intervening
        different authenticated request, an exact same-request retry returns
        the stored x1 after response loss; a consumed key_updated generation is
        not replayed.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferSenderRequest'
      responses:
        '200':
          description: Transfer secret share
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferSenderResponse'
        '400':
          $ref: '#/components/responses/MessageError'
        '401':
          $ref: '#/components/responses/MessageError'
        '404':
          $ref: '#/components/responses/MessageError'
        '500':
          $ref: '#/components/responses/MessageError'
  /transfer/update_msg:
    post:
      tags: [Transfer]
      summary: Store the encrypted transfer message
      description: |
        BIP448 requires the exact x1 public generation key and a
        domain-separated signature binding the statechain, recipient,
        generation, and hash of the decoded ciphertext.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferUpdateMsgRequest'
      responses:
        '200':
          description: Message updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedResponse'
        '401':
          $ref: '#/components/responses/MessageError'
        '404':
          $ref: '#/components/responses/MessageError'
        '500':
          $ref: '#/components/responses/MessageError'
...
  /withdraw/complete:
    post:
      tags: [Withdrawal]
      summary: Delete server rows after a cooperative withdrawal broadcast
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawCompleteRequest'
      responses:
        '200':
          description: Statechain rows deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '500':
          $ref: '#/components/responses/MessageError'
...
components:
...
  responses:
    MessageError:
      description: Request, upstream, or internal error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
  schemas:
...
    TransferSenderRequest:
      type: object
      required: [statechain_id, auth_sig, new_user_auth_key]
      properties:
        statechain_id:
          type: string
        auth_sig:
          type: string
        new_user_auth_key:
          type: string
        batch_id:
          type: string
          nullable: true
    TransferSenderResponse:
      type: object
      required: [x1]
      properties:
        x1:
          type: string
    TransferUpdateMsgRequest:
      type: object
      required: [statechain_id, auth_sig, new_user_auth_key, x1_pub, enc_transfer_msg]
      properties:
        statechain_id:
          type: string
        auth_sig:
          type: string
          description: |
            BIP448 current-owner signature over SHA256(domain ||
            u32_be(statechain_id_utf8_len) || statechain_id_utf8 ||
            recipient_key_compressed || x1_pub_compressed ||
            SHA256(decoded_ciphertext_bytes)), where domain is
            "BIP448/transfer-update-msg/v1\0".
        new_user_auth_key:
          type: string
        x1_pub:
          type: string
          description: |
            Canonical compressed public key derived from this transfer row's
            x1.
        enc_transfer_msg:
          type: string
    UpdatedResponse:
      type: object
      required: [updated]
      properties:
        updated:
          type: boolean
...
    WithdrawCompleteRequest:
      type: object
      required: [statechain_id, signed_statechain_id]
      properties:
        statechain_id:
          type: string
        signed_statechain_id:
          type: string
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string